Added Console Query Logger and Classes organized in separate files
