+++ Graphir.API/Schema/AttachmentType.cs
@@ -1,0 +1,18 @@
+﻿namespace Graphir.API.Schema;
+
+public class AttachmentType : ObjectType<Attachment>
+{
+    protected override void Configure(IObjectTypeDescriptor<Attachment> descriptor)
+    {
+        descriptor.BindFieldsExplicitly();
+
+        descriptor.Field(a => a.ContentType);
+        descriptor.Field(a => a.Language);
+        descriptor.Field(a => a.Data);
+        descriptor.Field(a => a.Url);
+        descriptor.Field(a => a.Size);
+        descriptor.Field(a => a.Hash);
+        descriptor.Field(a => a.Title);
+        descriptor.Field(a => a.Creation);
+    }
+}
