+++ Graphir.API/Schema/IResourceDelete.cs
@@ -1,0 +1,7 @@
+﻿namespace Graphir.API.Schema;
+
+[InterfaceType("ResourceDelete")]
+public interface IResourceDelete<T> where T : Resource
+{
+    public OperationOutcome Information { get; set; }
+}
