+++ Graphir.API/Schema/PatientCommunicationInput.cs
@@ -1,0 +1,6 @@
+﻿namespace Graphir.API.Schema;
+
+public record PatientCommunicationInput(
+    CodeableConceptInput? Language,
+    bool? Preferred
+);
