+++ Graphir.API/Schema/PatientContactInput.cs
@@ -1,0 +1,10 @@
+﻿namespace Graphir.API.Schema;
+
+public record PatientContactInput(
+    CodeableConceptInput[]? Relationship,
+    HumanNameInput? Name,
+    ContactPointInput[]? Telecom,
+    AddressInput? Address,
+    string? Gender,
+    PeriodInput? Period
+);
