+++ Graphir.API/Schema/PatientCreation.cs
@@ -1,0 +1,8 @@
+﻿namespace Graphir.API.Schema;
+
+public class PatientCreation : IResourceCreation<Patient>
+{
+    public string Location { get; set; }
+    public Patient Resource { get; set; }
+    public OperationOutcome Information { get; set; }
+}
