+++ Graphir.API/Schema/PatientDelete.cs
@@ -1,0 +1,6 @@
+﻿namespace Graphir.API.Schema;
+
+public class PatientDelete : IResourceDelete<Patient>
+{
+    public OperationOutcome Information { get; set; }
+}
