+++ Graphir.API/Schema/PatientUpdate.cs
@@ -1,0 +1,7 @@
+﻿namespace Graphir.API.Schema;
+
+public class PatientUpdate : IResourceUpdate<Patient>
+{
+    public Patient Resource { get; set; }
+    public OperationOutcome Information { get; set; }
+}
