+++ Graphir.API/Services/FhirAuthenticatedHttpMessageHandler.cs
@@ -13,8 +13,8 @@ public class FhirAuthenticatedHttpMessageHandler : DelegatingHandler
 
     protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
     {
-        string token = await _tokenService.GetAccessTokenForUserAsync(new[] { _fhirOptions.Scopes });
-
+        string token = await _tokenService.GetAccessTokenForAppAsync(_fhirOptions.Scopes);
+        
         request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", token);
 
         return await base.SendAsync(request, cancellationToken);
