+++ Graphir.API/appsettings.json
@@ -7,7 +7,7 @@
   },
   "AzureAd": {
     "Instance": "https://login.microsoftonline.com/",
-    "Domain": "phcc.gov.qa",
+    "Domain": "domain",
     "TenantId": "tenantId",
     "ClientId": "clientId",
     "ClientSecret": "clientSecret"
@@ -16,5 +16,10 @@
     "BaseUrl": "baseUrl",
     "Scopes": "scopes"
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "AppSettings": {
+    "Development": true,
+    "WebApiEnabled": true,
+    "EnableCachingInServiceProvider": true
+  }
 }
